Add a more unit tests for magnitude
For various reasons I was browsing the expression library code today,
and noticed that inside of the M() and/or ABS() operators, there are two
paths thru the code.    One is if the argument is a pure number, in
which case the std::abs function is used directly.

The other is if the argument is not a pure number (i.e. a parameter, or
another expression, etc).  Then it invokes the absOp operator class
instead.  Looking over the unit tests, I didn't see this use case tested
directly so I've added it.  It was tested indirectly, however so this
might be overkill, but it doesn't hurt.