Add suport for internal single quotes
The expression library has historically allowed internal fields to be
surrounded by curly braces.  It mostly treats them as a no-op.  In other
words:   {exp} = exp.

It occured to me today that since single quotes are also supported as
valid expression delimeters by the netlist parser, they should be
allowed inside of expressions.  The single quote token didn't appear to
be used in any other important way by the expression parser, so now,
inside of expressions, if you have 'exp' it is handled as 'exp' = exp.

This is kind of convenient if you are cutting/pasting things from the
parts of the netlist that are in the netlist parser domain, into
expressions, which are in the expression parser domain.  It is one less
thing to edit.