Refactor and performance improvement for tables
This is a minor refactor and performance enhancement.

Today I worked with a circuit with a very large PWL source.
In the original circuit the PWL source was a Vsrc.  However,
as an experiment to see what would happen, I changed the
Vsrc to a Bsrc, and made the syntax changes necessary
for it to be specified as a "TABLE".

These should be equivalent, although they use completely
different parts of the Xyce source code.   And I've done
experiments like this before, and this usually holds up.

However, today it was different.
Much to my surprise, after I switched from Vsrc to Bsrc,
the circuit ran much slower.  The reason, it turned out,
is that most of the value entries in the table were
specified via .params.  And, even though the values of the
.params didn't change (ie, they were not subject to a .step
sweep or anything like that), the table operator was
re-evaluating all of them every time the "evaluate"
function was called.

There was a little bit of logic in the "val()" function to
ensure that if the table was purely numerical values,
it wouldn't keep evaluating the subordinate expressions.
But, this logic wasn't smart enough.  I've updated it to
also account for constant .params.  This refactor isn't
perfect, but it is a significant improvement.

In my simple test circuit, which had thousands of PWL points
and a handful of resistors previously took several minutes to run
in the Bsrc version.  Now it take about 12 seconds on my laptop.
The Vsrc version takes a similar, but slightly slower amount
of time.

Also, in this commit, I noticed that I had an identical block
of code being executed in 4 places in the tableOp class.
I've replaced that with a single function.

As part of the refactor, the virtual AST function
"getIsConstant", which was originally used in a very limited
way is now used a lot more.  As such was necessary to
implement it in many more of the AST node classes, particular
the ones for simple math operations.  So that is in this
commit as well.