Update to correctly document LIMIT(x,y) and M(x)
M(x) was correctly documented, but it made sense to include it in the
complex operator section.

LIMIT(x,y,z)  (which is compatible with Pspice) was documented.

However, LIMIT(x,y), which is a random operator from Hspice, among
other codes, was not.   This 2-argument version got added about a year
ago and the issue was closed.  So, it not being in this table was a
mistake.

Xyce supports both, so they are now both in the table.