More updates to tables.
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/575
"Expression tables are inefficient when they are large and contain
expressions"

This commit creates a new function in the expression library, called
"updateForStep".  This work was movitivated by improving the speed of
the table operator, particularly for large tables which contain large
numbers of expressions.

Up until now, there have been 2 main functions in most AST nodes for
setting up expressions.  One has been the constructor (and subordinate
functions to the constructor) and the other has been the evaluations
functions such as val(), dx(), and dx2().

The problem is that some updates are expensive, but potentially need to
be done more than once, due to .param values changing.  But for
expensive updates, it is desirable to only update things when absolutely
necessary.  Functions like val(), dx(), and/or dx2() are called
a lot.  For Bsrc expressions, they are called every Newton iteration.

I had written the table operator to use std::vector<double> containers
for the table values.  These objects are  used by the various
interpolation functions.  But the table is (potentially) a bunch of
expressions, so each entry is also an abstract syntax tree (AST).  If
the AST depends on a changing .param it must be re-evaluated in order to
update the entries of the original table.  And, for more complicated
interpolations, such as splines, then the original setup functions for
the splines must be re-called as well.  On the other hand, if the
entries have not change, then all this work can be skipped at the only
work necessary for val() is the interpolation itself, which is cheap.

Since .param values don't (usually) change from time step to time step,
and, if they change at all, only change with .STEP iterations, it made
sense to create a new set of functions that are only called under those
circumstances.  This commit sets that up.  For now it is only used in
the table operator, but it could also be applied to other
expensive-to-set-up operators.

This commit also attempts to implement some logic for determining if a
subordinate AST is a constant expression or not.  Currently, I am only
using this logic in the new tableOp::updateForStep function, but I can
see lots of places in the expression library where this might be useful
in the future.