Fix for complex valued expressions on .PRINT line
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/28
"Make .PRINT expressions produce both real and imaginary part
automatically for frequency domain calculations"

I originally worked on this issue several years ago, found a quick fix,
but then rolled it back because that solution had some quirks I didn't
like.

For frequency domain analysis, the default behavior for most .PRINT
outputs is to automatically output the real and imaginary part of the
output variable.  So, for example, if the netlist has:

.print AC V(B)

then the output will be:   Re(V(B))  Im(V(B))

However, for expressions, this automatic output was never set up.  So,
if the equivalent expression was requested on the .PRINT line:

.print AC {V(B)}

then the output would be:  {V(B)}, which is just the real part.

Hacking in a quick fix was easy, but I didn't like that this would be
applied to any output, even ones that were purely real-valued.  For
example, the phase of a number, or the output noise, etc, would be real
valued even if the inputs were complex.  There were workarounds (in the
expression library the user can simply add Re() and/or Img() operators
to get the real and imaginary parts of any expression), so I put it
aside.

Today I have (again) fixed the issue.  The one thing that is different
this time is that the expression library now automatically determines if
an expression is likely to be real-valued or not.  In most cases, an
expression in AC analysis will probably have a complex result, but not
always.  The outputter now queries the expression to determine if it is
complex-valued or not, and then adjusts the outputs accordingly.

This commit also adds several new expression unit tests to test out the
"getIsComplex" function.  These tests are not comprehensive yet, so more
will come later.