Fix so that complex-valued .params can be used
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/29
"Come up with a comprehensive way of handling complex-valued parameters"

The old expression library didn't handle complex numbers, which meant
that all of Xyce's interactions with the expression library implicitly
assumed real-valued inputs and outputs to expressions.

The new expression library (now from a few years ago), was written to
support complex numbers, but a lot of the Xyce source code still
implicitly assumed that expressions and/or parameters were always
real-valued.   So, ever since the new expression library was integrated
into Xyce, I have periodically cleaned things up to support complex in a
more comprehensive manner.

One significant gap was the handling of .params.  Prior to this commit,
it was theoretically possible to specify a complex-valued parameter, but
due to the hardwired assumptions in the parser, by the time that
parameter was actually used by Xyce, it had been converted, and the
imaginary part was set to zero.  This conversion wasn't necessary, since
the expression library can support complex now.

This commit adds support for complex data types in the Xyce::Util::Param
classes, and also extends the code in various parts of the parser and
output "Op" classes to use these new data types.

So, before this commit, this wouldn't work correctly:

.param param1={1.0+2.0J}

.print AC {param1}

In the resulting output file, you would get 1.0 for the real part, and
0.0 for the imaginary part.  And, as the .PRINT line attempt to
auto-detect if "param1" is real or complex, it would probably just
print the real part.

With this commit, the above netlist fragment works correctly and will
correct print 1.0 and 2.0 for the real and imaginary parts,
respectively.  It also works correctly if the .param line doesn't have
curly braces or single quotes.