Fix failing parallel tests
In my commit from earlier today, I made a mistake that
caused Xyce to break in parallel with memory faults.

The commit earlier today was related to this work:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/29
"Come up with a comprehensive way of handling complex-valued parameters"

The mistake was in the N_UTL_Param.C file.  I had added support for
complex numbers to this class, which involved (among other things)
adding CMPLX and CMPLX_VEC to the enumeriated list of supported data
types.  Then, throughout the class, added code to support those
new data types.

The mistake I made was in the Param::unpack function, which, along with
the Param::packedByteCount and Param::pack functions are only used in
parallel.  Those functions contain switch statements based on the
data type.  When I added new sections to those switch statements to
support CMPLX and CMPLX_VEC, I inadvertantly deleted one of the "break;"
statements.  This caused the CMPLX_VEC unpack operations to be executed
when the should not have been, thus causing the memory error.

This is fixed now.