Interim commit related to Bsrc derivative handling
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/39
"Xyce fails regression tests on Windows when compiled with debug flags"

Recently, it was discovered that at least one regression test was
failing due to an indexing mistake with Bsrc derivatives.  The issue was
caused by an inconsistency in the total number of derivatives, and their
storage in a vector.

This inconsistency happened if both of these conditions were true:

1. The expression was from a Bsrc inside of a subcircuit.
2. The expression depended on subcircuit nodes (which are aliases)
3. The total number of unique nodes changed after aliasing, due to some
   of the subcircuit nodes being the same, or being ground.

For expressions that aren't based on aliased voltage nodes, the
uniqueness issue is easily handled during expression parsing.  Aliased
nodes are handled after expression parsing is over, however, and
requires that the relevant data structures be pruned.
The expression library was pruning some data structures but not
others.

So, that needs to be fixed.  However, it is not (yet) fixed in this
commit.  Before I could fix it, I needed to clean up the code a bit.

The device package handling of dependent parameters has been confusing
for various reasons.  One bit of confusion had to do with the handling
of "vars" coming from an expression.

In the DeviceEntity::setDependentParameter function, the following
things get set up for a device parameter that depends on an expression:

int dependentParam.lo_var
int dependentParam.n_vars
std::vector<std::string> expVarNames
std::vector<std::string> expVarTypes

where n_vars is the size of the expVarNames and expVarTypes vectors.

As the code was written, expVarNames would potentially contain the
names of voltage nodes, lead currents and device currents.  And, these
names were then used in the DeviceInstance:: getDepSolnVars function
call.  Basically, these should be a vector of soution variable names
that a Bsrc would depend on, from an expression.

The potential inclusion of lead currents in this vector made no sense,
as lead currents are not solution variables.   And, in practice, this
inclusion was inside of various conditional statements that were
impossible to reach, as far as I could tell.

I couldn't think of a legitmate reason for lead currents to be in this
vector, but on the off chance that there was one, I removed this
block of code and ran a LOT of tests.  Nothing broke.  So, this
commit removes lead currents from expVarNames and expVarTypes.

Once this was removed, then the expVarNames was unambiguously names
associated with expression/Bsrc derivatives.  To help prevent
future confusion, I changed the names of a few variables as well.

n_vars is now numVars
lo_var is now lowVarIndex

Also in this commit I've started testing functions related to derivative
setup and node aliasing in unit tests.