Remove inappropriately committted file
In 2016, in commit dc981e89, I imported the unmodified MEXTRAM code
into Xyce's git repo.

When I did so, I inappropriately committed a bootstrap file that was
created so that I could experiment with MEXTRAM as a plugin before
making it a real part of Xyce.

It should never have been committed to the repo.  I noticed today when
I pulled the public repo that it had been changed, as its copyright
date had been changed.  Since it should never have been there in the
first place, I'm just removing it while I still have access to do so.

Byeeeeeee.