Remove another bootstrap file
In commit b4425cca (2015) I had begun experimenting with VBIC 1.3, and
in doing so was building it as a plugin.  Plugins require "bootstrap"
files as part of the shared library, and I had added one to the
repo (which was still CVS at the time).

This unused file still exists in git, and it is a relic.

Removing it while I still have the chance.