Fix for related to subckt-Bsrc derivative handling
The original issue was this:
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/39
"Xyce fails regression tests on Windows when compiled with debug flags"

And the new, more specific, issue is this one:
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/41
"Expression library inconsistently handles subcircuit nodes"

This commit fixes the problem, in which the number of solution
derivatives from the expression library was inconsistent.

The inconsistency happened when:

1. The expression was from a Bsrc inside of a subcircuit.
2. The expression depended on subcircuit nodes (which are aliases)
3. The total number of unique nodes changed after aliasing, due to some
   of the subcircuit nodes being the same, or being ground.

The problem was that the expression library uses 3 containers to track
voltage nodes.  One is a vector of voltOp AST nodes.  I don't currently
have an easy way to swap AST nodes, so this vector always remains the
same.  Another is a map, which I use to track redundant nodes, and yet
another is a std::vector of strings, which contain the names of the
voltage nodes, in the same order as the voltOp vector.

This std::vector<std::string> object (voltNameVec_) is what the device
entity, and (ultimate) topology uses to generate a vector of LIDs, that
can be used to access relevant entries in the solution vector.  So, this
order needs to be consistent.

I had attempted to properly update voltOpMap and voltNameVec in the
"replaceNames" function, but I didn't get it quite right.  When a new
voltage node name resulted in a duplicate, the map gets updated
correctly, but the voltNameVec_ object did not.

One brute force fix is to force the function "setupVariousAstArrays" to
be executed.  When this happens, the voltOpMap and voltNameVec are both
regenerated, based on an AST traversal.  I try to avoid calling this
function, as it sets up everything.  But that is the solution in this
commit.

I'll note that this commit also  explicitly handles the ground node, to
exclude it from the Map and the names vector.   Unfortunately, to make
sure this is consistent, I have to make sure to use this "exclude
ground" logic everywhere the voltage operators are looped.  So, all the
group objects have been updated to account for it.