Fix index out of bounds events in UtilityPkg
The get from operators in the Marshal class performs an invalid memory
access if the container it’s reading from is empty. This results in
undefined behavior. This commit puts a check on the container size
around the read operation.

In the Util word_wrap method, there is an iterator that, in some
cases, is incremented past the end of its container. This commit
also adds a check that the iterator is not at the end before
incrementing it.

This issue was discovered in
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/39

This issue for fixing this specific problem
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/42