Fix Windows path check
The function changed in this commit is used to verify that a string
consists of a well-formed absolute file path. For Windows, the function
checks that the first slash appears after the disk drive indicator, but
the expected index is wrong. This change updates that check to use the
correct index.

This issue was discovered in
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/39

This issue for fixing this specific problem
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/43