Use vector data() method to get array pointer
There are places in the code where we need to get a pointer to the
underlying array of a vector. The &vector[0] form that is frequently
used to do this results in undefined behavior if the vector is empty.
There were a few places where this was happening. If the size of the
vector is not checked before attempting to access it, this commit
updates the code to use the data() method instead of the
index/subscript operator.

This commit does not attempt to replace all instances of &vector[0]
with vector.data(); just the ones that where found to be invoked with
empty vectors.

This issue was discovered in
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/39

This issue for fixing this specific problem
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/44