Fix index out of bounds errors
This commit corrects a couple of conditional checks on vector index
variables used inside of loops. In this set of corrections, the index
variable is set to a values outside of the range of the associated
vector. The code has been updated so that the index value is checked
before it is used to access the vector.

This issue was discovered in
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/39

This issue for fixing this specific problem
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/45