Fix for subckt parameter handling
This is a fix for this issue:

https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/54
"subcircuit parameters incorrectly deleted when multiple subcircuit
instances (X lines) refer to same subcircuit definition"

This issue was reported by an external user.  When subcircuits pass in
parameters, there was a scenario where the parameters passed into one
instance of a given subcircuit would incorrectly delete that parameter
from the subcircuit definition.  Then, for subsequent invocations of
that subcircuit, it would be unable to find that parameter.

This is a mistake I introduced relatively recently, when I refactored
the handling of subcircuit parameters.

In the Xyce parser, all subcircuits execute a "resolve" function, in
which all the subcircuit parameters for a given instances are
resolved.  If they depend on other parameters, etc, this is where those
dependencies are figured out.

This is done both for parameters defined inside the subcircuit and also
for parameters that are passed in via instance arguments.  As they get
resolved, they are removed from an "unresolvedParams" object and pushed
into a "resolvedParams" object.

The mistake that I made is that the "unresolvedParams" object is
associated with the subcircuit definition.   The resolve function uses a
local copy of this object when it does its work on a specific instance.
When processing the instance parameter arguments, I was erasing entries
from the definition object, rather than from the instance copied object.

This was easy to fix.  It is fixed in this commit.