Fixes device connectivity map for VCVS and VCCS devices
The device connectivity map that groups the leads/terminals of a
device that are connected was not implemented for these controlled
voltage sources.  This could result in errors in the topology checks
that use this information to determine that all nodes have a path
to ground.  This has been fixed in the commit.

https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/50
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/51