Delete dead code from CircuitBlock::addOptions
This is related to this issue:

"Modify Xyce parser so that it can handle multiple .options statements
for the same component"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/571

Fixing this issue will involve modifying the CircuitBlock::addOptions
function.  As a first start, I noticed a block of code in this function
that seemed to be a remnant of a much earlier parser design, that was no
longer relevant or used.  I commented it out and ran the test suite, and
observed no problems.  So, in this commit I am removing it.