Report failed gmin and source value on failed DC Op.
When diagnostic reporting is enable, a failed GMIN
or Source Stepping will notify the analysis manager
of the failed step and report the last GMIN
or source stepping value used.  This is put into
the diagnost output file for the user.

The diagnostic output file name now defaults to
<circuitname>.dia  Originally I had this default
to XyceDiag.out but as I used it more it became
obvious that it should behave like all of Xyce's
other output files and generate a name based
on the input file name.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/574