Allow brackets in device names in expressions
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/595
"Expression library does not recognize device names that include square brackets"

The netlist parser in Xyce allows the left and right bracket
characters in device names.  However, the expression parser
(which in theory should be consistent) did not allow this.
As a result, and expression like {I(V_B[0])} would fail.

This just required a small change to the regex for the "DEV" token in
the expression lexer.  This commit fixes the issue.