Visitor pattern refactor expression library
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/596
"Implement visitor pattern in the expression library"

One of the operations that has to be performed in the expression library
(especially once extern dependencies are resolved) is to traverse the
AST and gather information.

A common design pattern that is often used for this is the "visitor"
pattern.  I did not do the visitor pattern in the initial
implementation, partly because it is tricky to make it work with
ref-counted pointers.  The textbook implementation of visitor relies a
lot on the "this" pointer. The teuchos RCP doesn't have way to easily
obtain the RCP of "this" from inside of a class, AFAIK.

The boost shared_ptr does have a built-in way to do this, using the
std::enabled_shared_from_this feature.   But like I said, I don't think
the Teuchos::RCP has the equivalent feature.  So, this presented a bit
of a challenge, but I think I made it work.  Working around this made
the code a bit uglier, unfortunately.

But, even with the ugliness, it is less ugly than what was there before,
and it should run a little faster, as it no longer has to do any
run-time type checking, or if-statements.  This will mainly affect the
setup times of the library, so from a user point of view, this will
speed up parsing and setup.