Allow .options diagnostic lines to not have additional params
The PkgOptionsMgr was written to not allow an options line
with just a package name as in ".options diagnostic"  But
when I made the diagnostic output file optional, such lines
were valid. The PkgOptionsMgr would issue a waring and say
the line was ignored when in fact it was not.  So I fixed
the logic in the PkgOptionsMgr to allow ".options diagnostic"
lines to be ok.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/574