Unit tests for bracket char in device names
I previously created one unit test for this use case.  This commit adds
3 more.  This is related to this issue:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/595
"Expression library does not recognize device names that include square brackets"

Also incidentally included in this commit is some cleanup, as well as
adding new file, ast_visitor.h to the list of files in the Makefile.am
file in the expression directory.  This is related to my recent work
on this:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/596
"Implement visitor pattern in the expression library"

For that work, I added the new file, ast_visitor.h.  However I did not
add it to the Makefile.am file, which is used by the autotools build
system.  The consequences of this omission were small; it didn't prevent
Xyce from being buildable via the autotools.  The only consequence was
that the dependence on ast_visitor.h wasn't included in the makefile
dependencies.  We are moving away from autotools and adopting Cmake,
so this isn't very important, but it was easy enough to add this.