Fix for unusual characters in dev names
This commit expands the list of allowed characters in device names
in the expression library.  This is related to this issue:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/595
"Expression library does not recognize device names that include square
brackets and a bunch of other unusual (but supported) characters"

This issue was originally only about square bracket characters, but I've
now expanded it to include various other unusual characters.  This stuff
has been supported in the netlist parser for a while. The fact that the
expression library didn't handle it correctly was an oversight.

After this commit, the regular expression for the DEV token in the
expression lexer is :

DEV [a-zA-Z][a-zA-Z0-9_:\$\-`~!@#%&_\+|<>\?\.\\/\^\*\[\]]*

This covers most unusual characters, and is very similar to the regex
used for voltage node names.