Move call to fftw_cleanup() to ~Simulator()
Calling fftw_cleanup() in ~N_UTL_FFTWInterface() is not safe
as it will reset the FFTW library and invalidate, but not
delete, any fftw_plan data held by other instances of
N_UTL_FFTWInterface.  The safest place to call fftw_cleanup()
is in the destructor, ~Simulator(), where the extra data
that FFTW has accumulated can be safely deleted.

https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/40