Make "downsamples" in table op more robust
This commit addresses this issue:
"Implement the ability to downsample data pulses"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/564

I previously added a feature to the expression library, to allow the
user to request that tables be pruned to a smaller number of points,
using splines and gradient-based grid refinement.

One option I implemented was for the user to be able to request that the
grid refinement be performed using a linear or log scale.

Today I started making some tests for this new feature, and I discovered
that it was easy to break it, particularly for the log option.

This commit makes it (1) work a bit better and (2) not seg fault (!).

The problem was in the "checkAndFixPulse" function, which is called once
or twice during the grid refinement.  For the interpolations to work
well, it is important that the pulse not include any zero-valued points,
and also important that it be monotonic, and have no repeated time
points.  This "checkAndFixPulse" function was designed to remove points
that exhibited these problems.

Unfortunatley, I made a mistake. In the parts of the function that were
supposed to remove zero points, I had written it to exclude non-positive
points, ie. any points where y[i] <= 0.0.  When using the log option, it
was easy to have points where logy[i] < 0.0, even if y[i] > 0.0.  For
one example, it removed all the points in the table.  This then led to a
seg fault in the interpolator.

So, this commit changes the check to y[i] != 0.0, and also adds some
error traps in the interpolation classes, so even if something like this
happens again, the seg fault won't happen.