Adds Xyce_config.h to the N_PDS_fwd.h file
Due to the addition of error handling to the expression library
in commit a316107b5e8, the Xyce_config.h header is not
being included before the N_PDS_fwd.h header.  This results in Xyce
incorrectly determining that the build is not parallel, even when
it is, and then complaining about the redefinition of MPI_Op.

The result is build failures across all parallel platforms.  This
will be resolved by including the Xyce_config.h in the N_PDS_fwd.h
file.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/564