Refactor child nodes in AST
This commit is a checkpoint commit and contains the "child" portion
of a fix for this issue:

"Refactor children and parent nodes in expression library AST"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/598

This supports this issue:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24
"Treat "local" and "global" variation for UQ methods in manner similar
to Hspice"

This commit has refactored how child nodes are handling in all the AST
classes of the expression library.  Before, child nodes were purely
local variables to each class.  Now, they are handled using a base class
std::vector, called "childrenAstNodes_".

This doens't change the functionality of the expression library, but it
makes it easier to manage child nodes in a more abstract way.  This will
make it easier to replace and/or eliminate nodes from the tree, which is
the method I want to use to address local vs. global variation.