Make minor changes to CMake build instructions
This is not time-sensitive, but is the accumulation of some small
changes I have made over time. In INSTALL.md, I have attempted to make
the sample CMake invocations a little clearer based on some feedback
from a couple of users. I also modified the AMD and Trilinos
configuration files based on the upcoming Trilinos 14 release. More will
have to be done, but I wanted to capture the current state of things.