Completion of child/parent AST refactor
"Refactor children and parent nodes in expression library AST"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/598

This supports this issue:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24
"Treat "local" and "global" variation for UQ methods in manner similar
to Hspice"

This commit has completed the "parent" portion of this refector.  It
also includes new functions for replacing AST nodes, which make use of
the new child/parent vectors.

At this point the new functions are not (yet) deployed in the parser
in this commit.  But they will be as part of the fix for local vs global
variations.