Add new unit tests for node replace function
This supports this issue:
"Refactor children and parent nodes in expression library AST"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/598

and this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24
"Treat "local" and "global" variation for UQ methods in manner similar
to Hspice"

This commit also includes changing a all the unit test names so that
they no longer include undscores.  It was recently pointed out to me
that for the google test framework, you aren't supposed to use
underscores in test names.  It didn't seem to cause any problems before
(I've been using tests with underscores for years), but it didn't hurt
to change it.