Update UQ operators to handle local varation
"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

When I originally implemented the Xyce UQ methods, as well as the random
expression operators such as AGAUSS and AUNIF, I didn't realize that
there was as standard convention for local vs. global variation.  As a
result, Xyce treated everything as a global variation.

Here is the convention.  If you do this:

.param resval=aunif(1000,400)
r_one 1 2 resval
r_two 2 3 resval
r_three 3 4 resval

Then all 3 resistors should get unique random values.  (ie, they exhibit
"local" variation).

On the other hand if you do this:

.param resval=aunif(1000,400)
.param my_resval=resval
r_one 1 2 my_resval
r_two 2 3 my_resval
r_three 3 4 my_resval

Then all 3 resistors will get the same random value (ie, they exhibit
"global" variation).  Prior to this commit, Xyce would treat both of the
above cases as the same, global.

For parameter resolution, the Xyce parser and expression library mostly
use an "attachment" strategy, where the AST of an external node is
simply attached to the relevant AST node of the expression that
depends on it.  This strategy was more efficient, but it lead to random
being non-unique.  ie, if you attach "resval" to 3 resistors, they are
all pointing to the same single AST.

To fix this issue I implemented a new method for replacing AST nodes
rather than attaching them.  Then, for device parameters, if they depend
on a .param that contains a random operator, a copy of the AST from the
.param is substituted into the relevant part of the device parameter
AST.  This fixes the issue.

It appears that none of the existing regression tests depended on this
issue, so that means (1) even though this is a new behavior, it doesn't
break anything and (2) I'll have to create a small number of tests.