Add local/global variation to ref guide
This commit is about this recently fixed issue:

"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

This commit removes text from the "incompatibilities with other
simulators" table.  Xyce not handling local variation was a known issue
that had been documented in that table.  Now that it is supported, this
table entry can be removed.