Free char* space allocated in FFTW
A call to fftw_export_wisdom_to_strin() returns a char*
that was allocated in FFTW but needs to be free'ed in Xyce.
There are two calls to this routine and thus there were
two memory leaks where it wasn't free'ed.  Now the char*
space is free'ed by Xyce. I tested this on linux with
several netlists that both use and don't use FFTW with
GCC under valgrind and didn't get any memory errors.  Hopefully
this will clear the valgrind failures from last night.

https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/40