Create base and beta Trilinos initial cache files
To better standardize the Trilinos builds, based on the Trilinos 14.0
release, I have created new initial cache files. There are now "base"
and "beta" files. The regular base files (with and without MPI) will
result in Trilinos builds that enables a typical full builds of Xyce.
The beta files enable extra features that are not regularly tested.
In this set of files, those are ShyLU and ShyLU-Basker. The files also
include some comments about how to enable certain third-party solvers.

Finally, I updated the CMake install instructions to refer to the "base"
files.