Fix the handling of transient convergence issue at the first step
The usage of number of failed steps and number of failed local
truncation error tests are mixed. Also, the minimum time step is
calculated based on the current time and machine precision. This
leads to the incorrect handling of the transient convergence issue
at the first step out of a DC op. It reported that Xyce reaches
maximum local error test failures. But Xyce did not check LTE for the
first step. This commit fixes the incorrect usage of number of failed
steps. It also adds the time information if Xyce cannot take a step
during transient simulation. The functions related to restart feature
are updated as well

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/147