Fix memory errors in expression library
This is related to my work on these two issues:

"Refactor children and parent nodes in expression library AST"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/598

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24
"Treat "local" and "global" variation for UQ methods in manner similar
to Hspice"

These two issues involved refactoring the abstract syntax tree (AST) of
the expression library.  The refactor involved making the child and
parent node relationships more abstract, as well as implementing the
"visitor pattern".

One aspect of the visitor pattern is that it relies on the  "*this"
pointer.  However, in the expression library AST, the design uses
ref-count-ptrs (RCPs) extensively, and there is no version of "this"
implemented for RCPs.

So, in order to do the equivalent, I had to modify the visitor design
to pass around "this" RCPs.  That was fine, but along the way I got
the (bad) idea to have each AST node class own a copy of the "this"
RCP.  I didn't need to do this; my visitor implementation passed
"this" in as a function argument, and it could be used entirely
as a locally-scoped function variable.

The problem with storing it as a class variable is that it then had
circular ownership.  ie, there was:

  RCP -> ptr -> node class -> thisRCP -|
   ^                                   |
   |___________________________________|

where "thisRCP" and "RCP" in the above chart are the same RCP.

This circular dependence confused the ref-count process of the RCP.  I
experimented with explicitly setting the "thisRCP" to Teuchos::null in
the node destructor.  But, ultimately, the node classes simply didn't
need to own a copy of an RCP to itself, so the cleaner way to fix this
is to remove it. That way the node class can continue to use the default
destructor, which I prefer.

This commit also contains a fix to the astcomp.h file.  While working on
this memory bug, I noticed that I had 2 redundant macros in the file.
They were completely 100% the same, except for some minor whitespace
differences.  I would have expected that the compiler would complain
about having 2 identical macros defined, but apparently it did not.
Anyway, this commit also removes one of these redundant macros.

This commit does NOT remove all the memory issues.  It removes the ones
that were associated with the visitor pattern and some other related
code.  That covered most of the valgrind failures that have been
observed in nightly testing, but not all of them.  There are still
memory problems with the "replaceParameterNode" function, which
is a new function I implemented for this work.